backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
bcrypt==4.2.1
PyJWT==2.9.0
httpx==0.27.2
